requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
openai==1.68.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
